pipeline/pyproject.toml
[project]
name = "dangerous-robot-pipeline"
version = "0.1.0"
description = "PydanticAI agents for dangerousrobot.org research automation"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai>=0.1.0,<1.0",
    "pydantic>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[project.scripts]
ingest = "ingestor.cli:main"
consistency-check = "consistency.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["common", "ingestor", "consistency"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
